operator/bgp: Rename bgpv2 package to bgp
We now have only one BGP Control Plane version, so there is
no need to keep the version in the package name.
