+++ operator/pkg/bgp/cell.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package bgpv2
+package bgp
 
 import (
 	"github.com/cilium/hive/cell"
